+++ src/components/NewChallengeModal.js
@@ -4,11 +4,11 @@ import React, {
   useRef,
   useContext,
   Fragment,
+  useCallback,
 } from "react";
 import { useTranslation } from "react-i18next";
 import Spinner from "./Spinner";
 import { cloneDeep } from "lodash";
-import { API_ENDPOINT_OPEN } from "../config";
 import { gameinfo, GameFactory, addResource } from "@abstractplay/gameslib";
 import { MeContext } from "../pages/Skeleton";
 import Modal from "./Modal";
@@ -32,6 +32,7 @@ function NewChallengeModal(props) {
   const [clockHard, clockHardSetter] = useState(false);
   const [rated, ratedSetter] = useState(true); // Rated game or not.
   const [standing, standingSetter] = useState(false); // Standing challenge or not.
+  const [standingCount, standingCountSetter] = useState(0);
   const [opponents, opponentsSetter] = useState([]);
   const [nonGroupVariants, nonGroupVariantsSetter] = useState({});
   const groupVariantsRef = useRef({});
@@ -42,49 +43,14 @@ function NewChallengeModal(props) {
   }, [i18n.language]);
 
   useEffect(() => {
-    async function fetchData() {
-      try {
-        var url = new URL(API_ENDPOINT_OPEN);
-        url.searchParams.append("query", "user_names");
-        const res = await fetch(url);
-        const result = await res.json();
-        usersSetter(result);
-      } catch (error) {
-        errorSetter(error);
-      }
-    }
-    if (show && !opponent && users === null) {
-      console.log("Fetching users");
-      fetchData();
+    if ( (props.users !== undefined) && (props.users !== null) && (Array.isArray(props.users)) && (props.users.length > 0) ) {
+        usersSetter(props.users);
     } else {
-      console.log("NOT Fetching users");
+        usersSetter(null);
     }
-  }, [show, opponent, users]);
+  }, [props, usersSetter]);
 
-  useEffect(() => {
-    groupVariantsRef.current = {};
-    nonGroupVariantsSetter({});
-    if (props.fixedMetaGame !== undefined) {
-      metaGameSetter(props.fixedMetaGame);
-      handleChangeGame(props.fixedMetaGame);
-    } else {
-      metaGameSetter(null);
-    }
-    if (props.opponent === undefined) {
-      playerCountSetter(-1);
-      opponentsSetter([]);
-    } else {
-      playerCountSetter(2);
-      opponentsSetter([props.opponent]);
-    }
-    errorSetter("");
-    clockStartSetter(72);
-    clockIncSetter(24);
-    clockMaxSetter(240);
-    clockHardSetter(false);
-  }, [show, props]);
-
-  const setPlayerCount = (cnt) => {
+  const setPlayerCount = useCallback((cnt) => {
     playerCountSetter(cnt);
     if (cnt === 2) {
       seatingSetter("random");
@@ -96,9 +62,9 @@ function NewChallengeModal(props) {
     if (cnt !== -1) {
       opponentsSetter(Array(cnt - 1).fill(""));
     }
-  };
+  }, [playerCountSetter, seatingSetter, ratedSetter, opponentsSetter]);
 
-  const handleChangeGame = (game) => {
+  const handleChangeGame = useCallback((game) => {
     groupVariantsRef.current = {};
     if (game === "") {
       metaGameSetter(null);
@@ -129,7 +95,33 @@ function NewChallengeModal(props) {
       seatingSetter("random");
     }
     errorSetter("");
-  };
+  }, [metaGameSetter, allvariantsSetter, nonGroupVariantsSetter, setPlayerCount, playerCountSetter, opponentsSetter]);
+
+  useEffect(() => {
+    groupVariantsRef.current = {};
+    nonGroupVariantsSetter({});
+    if (props.opponent === undefined) {
+      playerCountSetter(-1);
+      opponentsSetter([]);
+    } else {
+      playerCountSetter(2);
+      opponentsSetter([props.opponent]);
+    }
+    errorSetter("");
+    clockStartSetter(72);
+    clockIncSetter(24);
+    clockMaxSetter(240);
+    clockHardSetter(false);
+    if (props.fixedMetaGame !== undefined) {
+      metaGameSetter(props.fixedMetaGame);
+      handleChangeGame(props.fixedMetaGame);
+    } else {
+      metaGameSetter(null);
+    }
+    if (props.opponent !== undefined) {
+        opponentsSetter([props.opponent])
+    }
+  }, [show, props, handleChangeGame]);
 
   const handleChangePlayerCount = (cnt) => {
     setPlayerCount(parseInt(cnt));
@@ -255,6 +247,7 @@ function NewChallengeModal(props) {
       metaGame: metaGame,
       numPlayers: playerCount,
       standing: standing,
+      duration: standingCount,
       seating: seating,
       variants: variants,
       challengees: opponents,
@@ -264,6 +257,7 @@ function NewChallengeModal(props) {
       clockHard: clockHard,
       rated: rated,
     });
+    handleNewChallengeClose();
   };
 
   let games = [];
@@ -304,11 +298,11 @@ function NewChallengeModal(props) {
     }
     playercounts = info.playercounts;
   }
-  if (!(metaGame === null || nonGroupData.length === 0)) {
-    console.log("nonGroupData", nonGroupData);
-    console.log(nonGroupVariants);
-  }
-  console.log(opponents);
+//   if (!(metaGame === null || nonGroupData.length === 0)) {
+//     console.log("nonGroupData", nonGroupData);
+//     console.log(nonGroupVariants);
+//   }
+//   console.log(opponents);
   return (
     <Modal
       show={show}
@@ -433,6 +427,8 @@ function NewChallengeModal(props) {
                   type="radio"
                   name="challengeType"
                   value="open"
+                  readOnly={true}
+                  checked={standing}
                   onClick={() => standingSetter(true)}
                 />
                 {t("ChallengeTypeOpen")}
@@ -442,7 +438,8 @@ function NewChallengeModal(props) {
                   type="radio"
                   name="challengeType"
                   value="targeted"
-                  defaultChecked
+                  checked={! standing}
+                  readOnly={true}
                   onClick={() => standingSetter(false)}
                 />
                 {t("ChallengeTypeTargeted")}
@@ -455,8 +452,7 @@ function NewChallengeModal(props) {
             </p>
           </div>
         )}
-        {playerCount === -1 || standing
-          ? ""
+        {playerCount === -1 || standing ? ""
           : /* Opponents */
             opponents.map((o, i) => {
               return (
@@ -508,7 +504,17 @@ function NewChallengeModal(props) {
                   </div>
                 </div>
               );
-            })}
+            })
+        }
+        {! standing ? "" : ( (playerCount === -1) || (playerCount > 2) ) ? "" :
+            <div className="field">
+                <label className="label is-small" htmlFor="duration">{t("Duration")}</label>
+                <div className="control">
+                    <input className="input is-small" type="number" min={0} name="duration" placeholder="duration" style={{width: "50%"}} value={standingCount} onChange={(e) => standingCountSetter(parseInt(e.target.value, 10))} />
+                </div>
+                <p className="help">{standingCount === 0 ? t("DurationHelpPersistent") : t("DurationHelp", {count: standingCount})}</p>
+            </div>
+        }
         {groupData.length === 0 && nonGroupData.length === 0 ? (
           ""
         ) : (