+++ src/index.js
@@ -10,6 +10,7 @@ require("core-js/es/set");
 require("core-js/es/promise");
 require("core-js/es/object");
 
+
 const container = document.getElementById("root");
 const root = createRoot(container);
 root.render(<Skeleton />);
@@ -17,4 +18,4 @@ root.render(<Skeleton />);
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
+serviceWorker.register();