+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Make the moves table header "sticky"
 * Added star feature.
 * Now supports alternate displays for games.
+* Added `duration` property to standing challenges for 2-player games. 0 means indefinite, >0 means it will expire after that many uses.
+* Push notifications!
 
 ### Fixed
 
@@ -47,7 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-* Fixed completed games table to contain more information and be more useful
+* Fixed completed games table to contain more information and be more useful.
+* Your Turn and Opponent's Turn tables now moved to new format, now including "time left".
 
 ## [v1.0.0-beta]